+++ common/src/main/java/net/william278/husksync/data/Identifier.java
@@ -58,6 +58,7 @@ public class Identifier implements Comparable<Identifier> {
             Dependency.optional("inventory")
     );
     public static final Identifier SOLCARROT_FOODLIST = huskSync("solcarrot_foodlist", true);
+    public static final Identifier ULTIMINE_ABILITY = huskSync("ultimine_ability", true);
     public static final Identifier ATTRIBUTES = huskSync("attributes", true,
             Dependency.optional("inventory"),
             Dependency.optional("potion_effects")
@@ -174,7 +175,7 @@ public class Identifier implements Comparable<Identifier> {
     public static Map<String, Boolean> getConfigMap() {
         return Map.ofEntries(Stream.of(
                         INVENTORY, ENDER_CHEST, POTION_EFFECTS, ADVANCEMENTS, LOCATION, STATISTICS,
-                        HEALTH, HUNGER, ATTRIBUTES, EXPERIENCE, GAME_MODE, FLIGHT_STATUS, PERSISTENT_DATA, MOD_DATA, SOLCARROT_FOODLIST
+                        HEALTH, HUNGER, ATTRIBUTES, EXPERIENCE, GAME_MODE, FLIGHT_STATUS, PERSISTENT_DATA, MOD_DATA, SOLCARROT_FOODLIST, ULTIMINE_ABILITY
                 )
                 .map(Identifier::getConfigEntry)
                 .toArray(Map.Entry[]::new));